[TS] Disallow 'null' on required string fields
If a field of type string is 'required', do not create the
underlying code assuming it can be 'null'.

Define its default value to an empty string.

This matches with TS type definitions where one of its members is a
string that is not optional, which in flatbuffers is represented with
the 'required' keyword in the schema.